feat:new folder-structure adds; new files and folders with updated names
