+++ src/components/infoRow.tsx
