--- src/navigation/types.ts
@@ -1,3 +1,0 @@
-export interface ComponentId {
-    componentId: string
-}
