+++ src/screens/PadDetailsScreen/index.tsx
@@ -2,20 +2,19 @@ import React, {FC, useEffect, useState} from 'react'
 import {ActivityIndicator, Platform, ScrollView, Text, View} from 'react-native'
 import LinearGradient from 'react-native-linear-gradient'
 import {Layout, Navigation} from 'react-native-navigation'
-import {applyAlpha, color, randomColor} from '../util/colors'
-import {InfoRow} from '../components/info-rox'
-import {LaunchListCell} from '../components/launch-list-cell'
-import {EmptyState} from '../components/empty-state'
-import {Launch} from '../api/types'
-import {PADS_STACK} from '../navigation/navigation'
-import {usePad, useRecentLaunches} from '../api/use-space-x'
-import {LaunchDetailLayout} from './launch-details'
-import {styles} from './pad-details.styles'
-import {ComponentId} from '../navigation/types'
+import {applyAlpha, color, randomColor} from '../../util/colors'
+import {InfoRow} from '../../components/infoRow'
+import {LaunchListCell} from '../../components/launchListCell'
+import {EmptyState} from '../../components/emptyState'
+import {ComponentId, Launch} from '../../types'
+import {PADS_STACK} from '../../navigation/navigation'
+import {usePad, useRecentLaunches} from '../../api/useSpaceX'
+import {LaunchDetailLayout} from '../LaunchDeatailsScreen'
+import {styles} from './styles'
 import { useDispatch, useSelector } from 'react-redux'
-import { State } from '../store/reducer'
-import { addToFavorites, removeFromFavorites } from '../store/actions'
-import { FavoriteButton } from '../components/favorite-button'
+import { State } from '../../store/reducer'
+import { addToFavorites, removeFromFavorites } from '../../store/actions'
+import { FavoriteButton } from '../../components/favoriteButton'
 
 interface PadDetailsProps {
     siteId: string;
