+++ src/screens/PadDetailsScreen/styles.ts
@@ -1,23 +1,23 @@
 import {Dimensions, StyleSheet} from 'react-native'
-import {applyAlpha, color} from '../util/colors'
+import {applyAlpha, color} from '../../util/colors'
 
 const {height, width} = Dimensions.get('window')
 
 export const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        justifyContent: 'center',
-        alignItems: 'center'
-    },
     loading: {
         flex: 1,
         justifyContent: 'center',
         alignItems: 'center'
     },
-    image: {
+    container: {
+        flex: 1,
+        justifyContent: 'center',
+        alignItems: 'center'
+    },
+    gradient: {
         width,
         height: height * 0.6,
-        resizeMode: 'cover'
+        justifyContent: 'flex-end'
     },
     imageOverlay: {
         flex: 1,
@@ -47,6 +47,13 @@ export const styles = StyleSheet.create({
         borderRadius: 4,
         overflow: 'hidden'
     },
+    description: {
+        fontSize: 14,
+        fontWeight: '500',
+        color: color.shade700,
+        paddingTop: 24,
+        paddingHorizontal: 20
+    },
     infoBox: {
         flex: 1,
         marginTop: 24,
@@ -56,17 +63,12 @@ export const styles = StyleSheet.create({
         borderWidth: 1,
         padding: 12
     },
-    youtubeContainer: {
-        marginTop: 24,
-        paddingHorizontal: 20
-    },
-    imageListContainer: {
-        marginTop: 24,
-        marginHorizontal: 20
-    },
-    flickrImage: {
-        marginVertical: 6,
-        width: '100%',
-        height: 300
+    recentLaunches: {
+        fontSize: 18,
+        fontWeight: '600',
+        color: color.shade700,
+        marginLeft: 20,
+        paddingTop: 24,
+        paddingBottom: 16
     }
 })
