+++ src/screens/PadsListScreen/index.tsx
@@ -1,12 +1,12 @@
 import React, {VFC} from 'react'
 import {FlatList, SafeAreaView} from 'react-native'
 import {LayoutComponent, Navigation} from 'react-native-navigation'
-import {EmptyState} from '../components/empty-state'
-import {PadListCell} from '../components/pad-list-cell'
-import {Pad} from '../api/types'
-import {PADS_STACK} from '../navigation/navigation'
-import {usePadsPaginated} from '../api/use-space-x'
-import {PadDetailLayout} from './pad-details'
+import {EmptyState} from '../../components/emptyState'
+import {PadListCell} from '../../components/padListCell'
+import {Pad} from '../../types'
+import {PADS_STACK} from '../../navigation/navigation'
+import {usePadsPaginated} from '../../api/useSpaceX'
+import {PadDetailLayout} from '../PadDetailsScreen'
 
 const PAGE_SIZE = 10
